projeto-letterboxd/requirements.txt
django
Pillow
requests
python-decouple